requirements.txt
fal-client>=0.5.0
requests>=2.28.0
python-dotenv>=1.0.0
